requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.2.25
charset-normalizer==3.4.7
emoji==2.15.0
idna==3.11
pyaes==1.6.1
pyasn1==0.6.3
python-telegram-bot==13.15
pytz==2026.1.post1
requests==2.33.1
rsa==4.9.1
six==1.17.0
Telethon==1.43.1
tornado==6.1
tzlocal==5.3.1
urllib3==2.6.3
